+++ payment/src/main/java/com/example/payment/application/dto/PaymentRefundItemResult.java
@@ -1,11 +1,12 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.PaymentRefundItemStatus;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PaymentRefundItemResult(
         UUID orderItemId,
         PaymentRefundItemStatus status,
-        Long refundAmount
+        BigDecimal refundAmount
 ) {
 }
